organized file structure and put typography components in specific page components instead of entire page
